refactor(task-graph): extract Create*Workflow factories to WorkflowFactories.ts
Move the four Create*Workflow factory functions plus their type aliases
(CreateWorkflow, CreateLoopWorkflow, CreateEndLoopWorkflow,
CreateAdaptiveWorkflow, EndLoopWorkflow) out of Workflow.ts into a
sibling WorkflowFactories.ts. Re-export from common.ts. Update direct
relative-path importers (FallbackTask, GraphAsTask, MapTask, ReduceTask,
WhileTask) to pull these names from WorkflowFactories.ts.

The Workflow.prototype.group / endGroup assignments stay inline in
Workflow.ts (after the class declaration) rather than moving with the
factories. Static imports in ESM are hoisted and evaluated depth-first,
so a tail side-effect import would re-enter WorkflowFactories.ts before
the Workflow class binding initializes and trip a TDZ on Workflow.prototype.
Keeping the proto assignments inline keeps the cycle safe: factory
function bodies reference Workflow lazily (at call time), so the
WorkflowFactories.ts <-> Workflow.ts cycle never dereferences an
uninitialized binding.
