+++ packages/task-graph/src/common.ts
@@ -19,6 +19,7 @@ export * from "./task-graph/Conversions";
 export * from "./task-graph/GraphToWorkflowCode";
 export * from "./task-graph/IWorkflow";
 export * from "./task-graph/Workflow";
+export * from "./task-graph/WorkflowFactories";
 export * from "./task-graph/WorkflowPipe";
 
 export * from "./task-graph/TransformRegistry";
